+++ CHANGELOG.md
@@ -11,12 +11,16 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Adds
 
 - Adds a variant for `NYCDOE` to the `Logo` component.
+- Adds the `TagSet` component.
+- Adds the `AlphabetFilter` component.
+- Adds the `AudioPlayer` component.
 - Adds the `requiredLabelText` prop to the `TextInput` and `Label` component.
   This allows customizing the "(Required)" text for other languages.
 
 ### Updates
 
 - Updates the `hamburger` SVG for the `Icon` component to allow for color assignments.
+- Updates `ComponentWrapper` prop `descriptionText` type to `string | JSX.Element`.
 
 ## 1.1.2 (September 29, 2022)